+++ Sources/KeyboardKit/Callouts/CalloutContext.swift
@@ -9,7 +9,6 @@
 import Combine
 import SwiftUI
 
-
 /// This class has observable, callout-related state.
 ///
 /// The ``inputContext`` property is used for input callouts,